Final version 1.0
This merge puts all the files together.

The 6BB Coders group worked hard for the past 8 weeks to achieve the
result. The team is proud to deliver 3 robot solutions: the line
follower, the line maze resolver, and the maze resolver. The team has
always strived to provide the client with the best coding solutions.
The code was brought up to the highest standards the group could have
thought of. With the project assignment, the team has improved its
skills in working in pairs, sub-dividing the assignments within the
group, and in programming itself.

With all honour and respect,
    Group 6BB Coders

Signed-off-by: paul <112290055+2560p@users.noreply.github.com>
Signed-off-by: Iarina <112045565+iarinamd@users.noreply.github.com>
Signed-off-by: ZsomborHajzer <114413262+ZsomborHajzer@users.noreply.github.com>
Signed-off-by: beyza olmez <beyza.olmez@student.nhlstenden.com>
Signed-off-by: timothyadewale <adewale.adewale@student.nhlstenden.com>
Signed-off-by: Bivol Maksym <maksym.bivol@student.nhlstenden.com>